+++ src/main/java/com/example/subway/line/domain/LineRepository.java
@@ -1,4 +1,4 @@
-package com.example.subway.line.dao;
+package com.example.subway.line.domain;
 
 import com.example.subway.line.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
