+++ src/main/java/com/example/subway/station/domain/StationRepository.java
@@ -1,5 +1,6 @@
-package com.example.subway;
+package com.example.subway.station.domain;
 
+import com.example.subway.station.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station,Long> {
