+++ src/main/java/com/example/subway/station/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package com.example.subway;
+package com.example.subway.station.dto;
 
 public class StationResponse {
     private Long id;
