+++ src/test/java/com/example/subway/line/application/LineServiceTest.java
@@ -1,11 +1,10 @@
 package com.example.subway.line.application;
 
-import com.example.subway.line.dao.LineRepository;
+import com.example.subway.line.domain.LineRepository;
 import com.example.subway.line.domain.Line;
 import com.example.subway.line.dto.LineRequest;
 import com.example.subway.line.dto.LineResponse;
 import com.example.subway.line.exception.LineDuplicateException;
-import com.example.subway.line.exception.LineExceptionType;
 import com.example.subway.line.exception.LineNotExistedException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
